+++ crypto/ecdsa.h
@@ -108,16 +108,16 @@ void ecdsa_get_address_raw(const uint8_t *pub_key, uint32_t version,
                            HasherType hasher_pubkey, uint8_t *addr_raw);
 void ecdsa_get_address(const uint8_t *pub_key, uint32_t version,
                        HasherType hasher_pubkey, HasherType hasher_base58,
-                       char *addr, int addrsize);
+                       char *addr, size_t addrsize);
 void ecdsa_get_address_segwit_p2sh_raw(const uint8_t *pub_key, uint32_t version,
                                        HasherType hasher_pubkey,
                                        uint8_t *addr_raw);
 void ecdsa_get_address_segwit_p2sh(const uint8_t *pub_key, uint32_t version,
                                    HasherType hasher_pubkey,
                                    HasherType hasher_base58, char *addr,
-                                   int addrsize);
+                                   size_t addrsize);
 void ecdsa_get_wif(const uint8_t *priv_key, uint32_t version,
-                   HasherType hasher_base58, char *wif, int wifsize);
+                   HasherType hasher_base58, char *wif, size_t wifsize);
 
 int ecdsa_address_decode(const char *addr, uint32_t version,
                          HasherType hasher_base58, uint8_t *out);
