+++ legacy/firmware/coins.c
@@ -61,8 +61,8 @@ bool coinExtractAddressType(const CoinInfo *coin, const char *addr,
                             uint32_t *address_type) {
   if (!addr) return false;
   uint8_t addr_raw[MAX_ADDR_RAW_SIZE] = {0};
-  int len = base58_decode_check(addr, coin->curve->hasher_base58, addr_raw,
-                                MAX_ADDR_RAW_SIZE);
+  size_t len = base58_decode_check(addr, coin->curve->hasher_base58, addr_raw,
+                                   MAX_ADDR_RAW_SIZE);
   if (len >= 21) {
     return coinExtractAddressTypeRaw(coin, addr_raw, address_type);
   }
